bids-validator/pyproject.toml
[tool.poetry]
name = "bids-validator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
minio = "^7.1.7"
xmltodict = "^0.12.0"
pydantic = "^1.9.0"
python-dotenv = "0.19.1"
pilot-platform-common = "0.0.25"
python-json-logger = "^2.0.2"
psycopg2-binary = "2.9.3"
click = "8.1.2"
black = "22.3.0"
SQLAlchemy = "^1.4.36"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
